+++ app/components/post.jsx
@@ -11,7 +11,7 @@ const Post = ({ post }) => {
             <h3>{titulo}</h3>
             <p className="fecha">{formatearFecha(publishedAt)}</p>
             <p className="resumen">{contenido}</p>
-            <Link  className="enlace" to={`/posts/${url}`}>Leer Post</Link>
+            <Link  className="enlace" to={`/blog/${url}`}>Leer Post</Link>
         </div>
     </article>
   )
