+++ app/routes/posts.$postUrl.jsx
@@ -1,16 +1,6 @@
 import { useLoaderData } from "@remix-run/react"
 import { getPost } from "~/models/post.server"
 import { formatearFecha } from "~/utils/helpers"
-import styles from "~/styles/blog.css"
-
-export function links () {
-    return [
-        {
-            rel: "stylesheet",
-            href: styles
-        }
-    ]
-}
 
 export function meta ({ data }) {
     if (!data) {
@@ -53,7 +43,7 @@ const Post = () => {
   const { titulo, contenido, imagen, publishedAt } = post?.data[0]?.attributes
 
   return (
-    <article className="contenedor post mt-3">
+    <article className="post mt-3">
        <img className="imagen" src={imagen?.data?.attributes?.url} alt={`imgen blog ${titulo}`} />
         <div className="contenido">
             <h3>{titulo}</h3>
